add rudimentary screen-reader support
This adds tabindex values and labels to controls,
and adds screen-reader applicable line breaks to the output display.

Synthetic line breaks need to be added
because Screen-readers treat span elements as inline elements by default,
regardless of CSS.
There are a lot of changes that can be done here, but this at least allows for
playback and streaming of events.

The most urgent next fix here would be to
isolate the parts of lines that have changed, and just read out those changes,
using the aria-live role.

Change-Id: I362189958b8489cfd9abb3baac95e72be81a7a25