Place control bar below last terminal line
This fixes problem with the control bar obscuring last 1-2 terminal
lines during mouse movement. The fix keeps the control bar below
terminal. Control bar still gets auto hidden when there's no user
(mouse) activity, but now it disappears in place (via `opacity`).

The slight downside of this fix is that the space needed by the control
bar is now permanently occupied regardless if it's visible or hidden,
which for the most part looks like a little bit of extra padding on the
bottom. Still, from UX point of view I believe this is better.